price customization by swapping from float to DecimalField

+++ Ecommerce/store/models.py
@@ -12,7 +12,7 @@ class Customer(models.Model):
         return self.name
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
